+++ examples/native/include.am
@@ -1,0 +1,14 @@
+# vim:ft=automake
+# All paths should be given relative to the root
+
+if BUILD_EXAMPLES
+noinst_PROGRAMS += examples/native/native_test
+noinst_HEADERS  += examples/native/native_test.h
+examples_native_native_test_SOURCES      = examples/native/native_test.c \
+									       examples/tpm_io.c
+examples_native_native_test_LDADD        = src/libwolftpm.la $(LIB_STATIC_ADD)
+examples_native_native_test_DEPENDENCIES = src/libwolftpm.la
+endif
+
+dist_example_DATA+= examples/native/native_test.c
+DISTCLEANFILES+= examples/native/.libs/native_test
