--- examples/tpm/include.am
@@ -1,13 +1,0 @@
-# vim:ft=automake
-# All paths should be given relative to the root
-
-if BUILD_EXAMPLES
-noinst_PROGRAMS += examples/tpm/tpm2_demo
-noinst_HEADERS += examples/tpm/tpm2_demo.h
-examples_tpm_tpm2_demo_SOURCES      = examples/tpm/tpm2_demo.c
-examples_tpm_tpm2_demo_LDADD        = src/libwolftpm.la $(LIB_STATIC_ADD)
-examples_tpm_tpm2_demo_DEPENDENCIES = src/libwolftpm.la
-endif
-
-dist_example_DATA+= examples/tpm/tpm2_demo.c
-DISTCLEANFILES+= examples/tpm/.libs/tpm_demo
