+++ examples/wrap/include.am
@@ -1,0 +1,14 @@
+# vim:ft=automake
+# All paths should be given relative to the root
+
+if BUILD_EXAMPLES
+noinst_PROGRAMS += examples/wrap/wrap_test
+noinst_HEADERS  += examples/wrap/wrap_test.h
+examples_wrap_wrap_test_SOURCES      = examples/wrap/wrap_test.c \
+									   examples/tpm_io.c
+examples_wrap_wrap_test_LDADD        = src/libwolftpm.la $(LIB_STATIC_ADD)
+examples_wrap_wrap_test_DEPENDENCIES = src/libwolftpm.la
+endif
+
+dist_example_DATA+= examples/wrap/wrap_test.c
+DISTCLEANFILES+= examples/wrap/.libs/wrap_test
